tratamento erro 404

+++ alura/src/main/java/br/com/alura/forum/controllers/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.alura.forum.controllers;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -57,28 +58,39 @@ public class TopicosController {
 		URI uri = uriBuider.path("/topico/{id}").buildAndExpand(topico.getId()).toUri();
 		return ResponseEntity.created(uri).body(new TopicoDto(topico));	
 	}
+	
+	
 	@GetMapping("/{id}")
-	public TopicoDto detalhar(@PathVariable Long id) {
-		Topico topico = topicoRepository.getReferenceById(id);
-		return new TopicoDto(topico);
+	public  ResponseEntity <DetalhesDoTopicoDto> detalhar(@PathVariable Long id) {
+		Optional<Topico> topico = topicoRepository.findById(id);
+		if(topico.isPresent()) {
+			return ResponseEntity.ok( new DetalhesDoTopicoDto(topico.get()));
+		}
+		
+		return ResponseEntity.notFound().build();
 	}
 	
 	//put serve para reescrever o metodo inteiro
 	@PutMapping("/{id}") 
 	@Transactional
 	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form){
-		Topico topico = form.atualizar(id, topicoRepository);
-	
-		return ResponseEntity.ok(new TopicoDto(topico));
-	
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()) {
+			Topico topico = form.atualizar(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDto(topico));
+		}
+		
+		return ResponseEntity.notFound().build();
 	}
 	
 	@DeleteMapping("/{id}")
 	public ResponseEntity <?> remover(@PathVariable long id){
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()) {
 		topicoRepository.deleteById(id);
 		return ResponseEntity.ok().build() ;
-
-		
+		}
+		return ResponseEntity.notFound().build();
 	}
 
 }
